客户汇款记录接口已完成

+++ provider/provider-fiveP-api/src/main/java/com/funtl/myshop/plus/provider/dto/IncomeDto.java
@@ -21,5 +21,5 @@ public class IncomeDto implements Serializable {
     private Integer rows;
 
     @ApiModelProperty(value = "总汇款金额")
-    private double totalAmt;
+    private BigDecimal totalAmt;
 }
